# swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is the API for the backend server (orchestrator) of the ePrescription
    prototype for the ACIT Cloud Studio
  version: 1.0.0
  title: Swagger ePrescription
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: thomas.i.jacobs@accenture.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: e-prescription.mybluemix.net
basePath: /
...
schemes:
  - http
paths:
  /login:
    post:
      tags:
        - Login
      summary: Validate user's login
      description: >-
        This endpoint will validate the user_name/password from a local database
        and route the user to the correct home screen based on the type of
        participant they are (doctor, pharmacy, patient)
      operationId: login
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: credentials
          description: Login details for user
          required: true
          schema:
            $ref: '#/definitions/Login'
      responses:
        '200':
          description: User logged on successfully
          schema: 
            $ref: '#/definitions/LoginOutput'
        '405':
          description: User name not found
...
  /get-history:
    get:
      tags:
        - Blockchain
      summary: Gets transaction history for a particular asset
      description: >-
        This endpoint will take the prescriptionID to get the transaction history.
      operationId: gethistory
      produces:
        - application/json
      parameters:
        - in: query
          name: prescription_id
          description: prescriptionId to check to get history
          required: true
          type: string
      responses:
        '200':
          description: Got History.
          schema: 
            type: array
            items:
                $ref: '#/definitions/History_info'
        '405':
          description: Was not able to get history
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  Login:
    type: object
    required:
      - user_name
      - password
    properties:
      user_name:
        type: string
        format: Joe Patient
      password:
        type: string
        format: password
  LoginOutput:
    type: object
    properties:
      userID:
        type: number
      role:
        type: string
      name: 
        type: string
      location:
        type: string
  History_info:
    type: object
    properties:
        provider_id:
          type: string
        pharmacy_id:
          type: string
        timestamp:
          type: string
          format: date-time
        transaction:
          type: string
...